- python cli/main.py add commands to load to db - renamed verified_stub_generator to cli
