+++ cli/templates/catalog.py.txt
