+++ cli/templates/catalog.yml.txt
