+++ cli/templates/generator.py.txt
