+++ cli/templates/specs.yml.txt
