+++ cli/templates/streams.py.txt
