+++ cli/templates/tests/generator/conftest.py.txt
