+++ cli/templates/tests/generator/test_.py.txt
